exercise DPHI bootcamp
analysis of surviving patients after a series of drug treatments used in a hospital. In order to try to predict whether there is a relationship between the types of treatments and the survival condition of a patient, a mathematical model was created from random forest model after a exploratory data analysis.